fix: extract email validation logic
