+++ app/javascript/pages/account/components/EmailFieldset.tsx
@@ -8,6 +8,25 @@ const validateEmail = (email: string) => {
   return emailRegex.test(email)
 }
 
+const emailValidation = (data: string) => {
+  const numberOfAts = (data.match(/@/g) || []).length
+  if (numberOfAts === 0) {
+    return t("error.email.missingAtSign")
+  }
+
+  const splitString = data.split("@")
+  if (
+    splitString[splitString.length - 1] &&
+    splitString[splitString.length - 1]?.search(/\./) === -1
+  ) {
+    return t("error.email.missingDot")
+  }
+
+  if (!validateEmail(data)) {
+    return t("error.email.generalIncorrect")
+  }
+}
+
 interface EmailFieldProps {
   register: UseFormMethods["register"]
   defaultEmail?: string
@@ -26,24 +45,7 @@ const EmailFieldset = ({ register, errors, defaultEmail, onChange, note }: Email
         placeholder="example@web.com"
         validation={{
           required: true,
-          validate: (data: string) => {
-            const numberOfAts = (data.match(/@/g) || []).length
-            if (numberOfAts === 0) {
-              return t("error.email.missingAtSign")
-            }
-
-            const splitString = data.split("@")
-            if (
-              splitString[splitString.length - 1] &&
-              splitString[splitString.length - 1]?.search(/\./) === -1
-            ) {
-              return t("error.email.missingDot")
-            }
-
-            if (!validateEmail(data)) {
-              return t("error.email.generalIncorrect")
-            }
-          },
+          validate: emailValidation,
         }}
         error={errors.email}
         errorMessage={errors.email?.message}
